+++ modules/core/src/test/scala/org/tessellation/infrastructure/snapshot/GlobalSnapshotConsensusFunctionsSuite.scala
@@ -1,6 +1,6 @@
 package org.tessellation.infrastructure.snapshot
 
-import cats.data.{NonEmptyList, NonEmptySet}
+import cats.data.NonEmptyList
 import cats.effect.std.Supervisor
 import cats.effect.{IO, Resource}
 import cats.syntax.applicative._
@@ -10,7 +10,7 @@ import cats.syntax.list._
 import scala.collection.immutable.{SortedMap, SortedSet}
 
 import org.tessellation.currency.schema.currency.{CurrencyIncrementalSnapshot, CurrencySnapshotInfo, SnapshotFee}
-import org.tessellation.domain.rewards.Rewards
+import org.tessellation.domain.rewards.EpochRewards
 import org.tessellation.ext.cats.syntax.next._
 import org.tessellation.kryo.KryoSerializer
 import org.tessellation.schema._
@@ -22,7 +22,7 @@ import org.tessellation.schema.transaction.{DAGTransaction, RewardTransaction}
 import org.tessellation.sdk.config.AppEnvironment
 import org.tessellation.sdk.domain.block.processing._
 import org.tessellation.sdk.domain.snapshot.storage.SnapshotStorage
-import org.tessellation.sdk.infrastructure.consensus.trigger.EventTrigger
+import org.tessellation.sdk.infrastructure.consensus.trigger.{ConsensusTrigger, EventTrigger}
 import org.tessellation.sdk.infrastructure.metrics.Metrics
 import org.tessellation.sdk.infrastructure.snapshot.{GlobalSnapshotAcceptanceManager, GlobalSnapshotStateChannelEventsProcessor}
 import org.tessellation.sdk.sdkKryoRegistrar
@@ -101,22 +101,11 @@ object GlobalSnapshotConsensusFunctionsSuite extends MutableIOSuite with Checker
 
   val collateral: Amount = Amount.empty
 
-  val rewards: Rewards[IO] = new Rewards[IO] {
+  val rewards: EpochRewards[F] = new EpochRewards[IO] {
+    override def distribute(artifact: Signed[GlobalSnapshotArtifact], trigger: ConsensusTrigger): IO[SortedSet[RewardTransaction]] =
+      IO(SortedSet.empty)
 
-    override def mintedDistribution(
-      epochProgress: EpochProgress,
-      facilitators: NonEmptySet[ID.Id]
-    ): IO[SortedSet[RewardTransaction]] = ???
-
-    override def feeDistribution(
-      snapshotOrdinal: SnapshotOrdinal,
-      transactions: SortedSet[DAGTransaction],
-      facilitators: NonEmptySet[ID.Id]
-    ): IO[SortedSet[RewardTransaction]] = IO(SortedSet.empty)
-
-    override def getAmountByEpoch(epochProgress: EpochProgress, rewardsPerEpoch: SortedMap[EpochProgress, Amount]): Amount =
-      ???
-
+    override def getAmountByEpoch(epochProgress: EpochProgress, rewardsPerEpoch: SortedMap[EpochProgress, Amount]): Amount = Amount.empty
   }
 
   def mkGlobalSnapshotConsensusFunctions()(implicit ks: KryoSerializer[IO], sp: SecurityProvider[IO], m: Metrics[IO]) = {
